Move middleware/startup class to separate folder
