+++ Grand.Framework/StartupConfigure/UrlRewriteStartup.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System.IO;
 
-namespace Grand.Framework.Infrastructure
+namespace Grand.Framework.StartupConfigure
 {
     /// <summary>
     /// Represents object for the configuring/load url rewrite rules from external file on application startup
